+++ lib/log_utils_lib.dart
@@ -1,9 +1,8 @@
-library;
-
 import 'package:flutter/foundation.dart';
 import 'package:logger/logger.dart';
 
-export 'package:logger/src/log_level.dart';
+export 'package:logger/logger.dart'
+    show AnsiColor, DateTimeFormat, Level, LogOutput, LogPrinter, OutputEvent;
 
 part 'src/log_print.dart';
 part 'src/log_utils.dart';